create shower_thoughts
this is a pretty *cool* file. i'm not actually fully done with it (see the TODO comment) but it's going to take waayyy too long for me to figure that out. so, in the meantime, i'll run this. hopefully someone can please help me with the auth system for a thought, just so that it isn't sending explicit stuff to the channel.